chess_pit/backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "chess-pit-backend"
dynamic = ["version"]
description = "FastAPI backend service for Chess Pit."
authors = [{name = "Arcade Hub", email = "devs@arcadehub.local"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<0.116",
  "uvicorn[standard]>=0.30,<0.32",
  "sqlmodel>=0.0.21,<0.0.22",
  "pydantic-settings>=2.0,<3.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "passlib[bcrypt]>=1.7,<1.8",
  "python-multipart>=0.0.9,<0.0.10",
  "alembic>=1.13,<1.14"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<0.28",
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.23,<0.25",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.6.6,<0.7.0"
]

[tool.hatch.version]
path = "src/chess_backend/__init__.py"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.hatch.build]
include = [
  "src/**",
  "alembic.ini",
  "alembic/**",
]

[tool.hatch.build.targets.wheel]
packages = ["src/chess_backend"]

[tool.hatch.build.targets.sdist]
include = [
  "src/chess_backend",
  "alembic.ini",
  "alembic",
  "README.md",
]
